Add color support
